registry/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv # For optional auth credentials
itsdangerous  # For signed session cookies 